requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
onnx>=1.12.0
onnxruntime>=1.14.0